Sets up behaviours for onesize and balance
Currently a little uncertain as to how we are targeting / identifying these elements. The behaviour of the balance, or onesize attribute has an effect on the div.o-question-response which has a  data-questiongroup attribute that could be targeted. This would mean doing an app.registerComponent('oQuestion', '<groupid>') and having the app.RegisterProperties('<groupid>') match. Presently just adds colours to corresponding elements to indicate they are targeted.