Resolves an issue with the initial row/column/grand totals
Occasionally totals were not initialising correctly. This is partially due to an order-of-init issue but shouldn't have caused a problem here as grids need to scan their inputs as part of building the monitoring method and should, therefore, always have the data to hand. This resolves the issue by putting the initial 'recalculate' call in a sensible place - AFTER we have built the array of elements.