Resolve an issue where buttons and m-option-base do not balance
Note to self: m-option-base items, and indeed general elements, should stretch the full width of their parent container to ensure the following element appear below. When balancing this width collapses down to allow for multiple columns. Where buttons and input elements such as m-option-base co-exist (chose a button or not applicable checkbox) the 100% width of the m-option-base forces all the buttons to also become full width.