Improve initial droplist element sizing
There's an issue with resize events in droplists where they fire as items within the list finish initialising. This can result in the final resize being less than the size of the largest item. Taking steps to investigate and resolve but this is a temporary fix to ensure we don't end up with zero-width items.