Resolve initialisation issue with drop lists in grids
Requires HTML fix to be effective: o-question-response class must be removed from div with the class o-question-table, at present the outer div is causing the code from incorrectly identifying all children as belonging to the questiongroup that is assigned to it. This change also moves some init() from the constructor into a function that runs on the DOM being ready as Chrome was executing the JS faster than the DOM was being built and would on occasion look for elements that hadn't fully build (e.g. an m-option-base that didn't have a label).