Reverses component initialisation order
Changes the initialisation order so that the page component initialises first. This should be a general improvement as components will now begin initialisation in the order that makes most sense (begin in order the components are added) and means that component sizing is more likely to take place after any page side-by-side adjustments have been configured.