Corrects a ticklabel/mark problem
If the user has specified a tick label interval of 180 minutes and the min/max times are divisible into chunks of three hours the outcome is as expected. However, if the time span isn't neatly divisible by the tick label interval there is a disconnect between the position of the labels and the actual time on the slider.

For example I have a test page where one of the time ranges is from 10:59-23:59 with a desired label interval of 180 minutes. The closest even division of labels that can be divided into even chunks is 195 minutes. This leads to labels at 10:59, 14:14, 17:29, 20:44 and 23:59. This keeps consistency with the labels and honours the user's requested value as closely as possible but looks... weird.

It's better than a broken label though.