Allows components to request a value
Values are broadcast from components as they finish initialising or when their data changes. During initialisation it is occasionally necessary for components to know the value of other components on the page, specifically a thumb label may require the value from its slider. Previously this has been a tightly coupled affair but this change seeks to improve that by allowing components to send a request for values to other members of their group.