Flatten the component array
Initialised components are stored in an array of objects (app.components). I noted that each item in the array stored a sub-array which contained a single entry (the component). This removes the sub-array so that the app.components array is a flat array of component objects.

Before:
app.components: [
[0]: [0]: {object}
[1]: [0]: {object}
]

After:
app.components: [
[0]: {object}
[1]: {object}
]