Fix regression where list items could fail to size correctly.
In a desire to remove warning flags from my code I pushed a value to the calculate width function during initialisation. This meant that the 'did I get a value' check was always passing and the input sometimes would miss the fact that it hadn't received a width broadcast from the list.