Extend answerCount to include list questions
It was observed that answerCount rules were not working when a list-type question was being evaluated. Nothing in the code was specifically preventing this but lists were returning an empty text string when no selection had been made. The empty string was being evaluated as an answer (a empty text string in an array is still a countable entry). This update extends the checks on blank entries to prevent them from being inserted in arrays while allowing them in 'simple' evaluations such as '' <= 30.